image_process_app/pyproject.toml
[tool.poetry]
name = "image-process-app"
version = "0.1.0"
description = ""
authors = ["Daniil Maximov <mcsdanf98@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiokafka = "^0.12.0"
pillow = "^11.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
